Add a generic address decoder
The decoder is built like a multiplexer: depending on the number
of bits N received as input, it is implied that there are 2**N RAM
cells. The selected RAM cell (CS = 1) is the n-th one with n the
value of the input.

A testbench is added both in VHDL and with Cocotb to allow easier
testing, also allowing testing genericity.